+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #74 Add a @BQConfigProperty for 'compactPath' in ContextHandler
 * #75 Cleaning up APIs deprecated since <= 0.25
 * #76 Upgrade Jetty to 9.4
+* #77 Use percentage metric for thread pool utilization check
 
 ## 0.25